cleanups, use defined width integer types

+++ lib_aes67/src/media_clock_and_control/media_clock_client.h
@@ -34,8 +34,8 @@ int aes67_get_buf_ctl_adjust(chanend buf_ctl);
 int aes67_get_buf_ctl_cmd(chanend buf_ctl);
 void aes67_send_buf_ctl_info(chanend buf_ctl,
                              int active,
-                             unsigned int ptp_ts,
-                             unsigned int local_ts,
-                             unsigned int rdptr,
-                             unsigned int wrptr,
+                             uint32_t ptp_ts,
+                             uint32_t local_ts,
+                             uintptr_t rdptr,
+                             uintptr_t wrptr,
                              timer tmr);
